Backend/requirements.txt
flask
flask-cors
openai
librosa
soundfile
numpy
pyyaml
scipy